relocate lo36 to /b
