+++ app/b/c/[thread]/ChatPage.tsx
@@ -4,9 +4,9 @@ import * as Icons from '@phosphor-icons/react/dist/ssr'
 import { Button, ScrollArea } from '@radix-ui/themes'
 import { useRouter } from 'next/navigation'
 
-import { Composer } from '@/app/dev/lo36/c/[thread]/_components/composer/Composer'
-import { Message } from '@/app/dev/lo36/c/[thread]/_components/Message'
-import { ChatProvider, useChat } from '@/app/dev/lo36/c/[thread]/_provider/ChatProvider'
+import { Composer } from '@/app/b/c/[thread]/_components/composer/Composer'
+import { Message } from '@/app/b/c/[thread]/_components/Message'
+import { ChatProvider, useChat } from '@/app/b/c/[thread]/_provider/ChatProvider'
 import { Link } from '@/components/ui/Link'
 import { Skeleton } from '@/components/ui/Skeleton'
 import { cn } from '@/lib/utils'
