+++ app/b/c/[thread]/_components/Message.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import * as Icons from '@phosphor-icons/react/dist/ssr'
 import { DropdownMenu, IconButton } from '@radix-ui/themes'
 
-import { Marble } from '@/app/dev/lo36/c/[thread]/_components/Marble'
+import { Marble } from '@/app/b/c/[thread]/_components/Marble'
 import AudioPlayer from '@/components/audio/AudioPlayer'
 import { ImageCard } from '@/components/images/ImageCard'
 import { Markdown } from '@/components/message/Markdown'
