+++ app/b/c/[thread]/page.tsx
@@ -1,4 +1,4 @@
-import { ChatPage } from '@/app/dev/lo36/c/[thread]/ChatPage'
+import { ChatPage } from '@/app/b/c/[thread]/ChatPage'
 
 export default function ThreadPage({ params }: { params: { thread: string } }) {
   return <ChatPage slug={params.thread} />
