+++ app/b/config.ts
@@ -1,0 +1,4 @@
+export const appConfig = {
+  baseUrl: '/b',
+  chatUrl: '/b/c',
+} as const
