--- app/dev/lo36/config.ts
@@ -1,4 +1,0 @@
-export const appConfig = {
-  baseUrl: '/dev/lo36',
-  chatUrl: '/dev/lo36/c',
-} as const
