+++ components/layout/NavRail.tsx
@@ -18,7 +18,7 @@ export const NavRail = ({ className, ...props }: NavRailProps) => {
     <nav
       {...props}
       className={cn(
-        'h-12 w-full shrink-0 gap-4 p-2.5 flex-start sm:h-full sm:w-12 sm:py-3 sm:flex-col-start',
+        'flex-start h-12 w-full shrink-0 gap-4 p-2.5 sm:flex-col-start sm:h-full sm:w-12 sm:py-3',
         className,
       )}
     >
@@ -29,7 +29,7 @@ export const NavRail = ({ className, ...props }: NavRailProps) => {
 
       <AdminOnlyUi>
         <IconButton variant="ghost" className="shrink-0" asChild>
-          <Link href={`/dev/lo36/${path.slice(1)}`}>
+          <Link href={`/b/${path.slice(1)}`}>
             <Icons.Planet className="size-6" />
           </Link>
         </IconButton>
